Refactored code into classes
